photo_culling/requirements.txt
numpy>=1.24
Pillow>=9.0
scipy>=1.10
# Optional: for CLIP-based semantic similarity
# torch>=2.0
# transformers>=4.30
